changed state, passing data to chart
