+++ client/src/pages/UserPage/style.css
@@ -1,0 +1,60 @@
+.trip-card {
+    background-color: rgba(66, 66, 66, 0.74);
+    font-family: 'Pontano Sans', sans-serif;
+    color: rgb(255, 255, 255);
+}
+
+.trip-header {
+    font-size: 20px;
+    text-align: center;
+}
+
+.trip-table {
+    color: rgb(255, 255, 255);
+}
+
+th {
+    font-size: 18px;
+}
+
+td {
+    font-size: 17px;
+}
+
+.remove-button {
+    color: rgba(255, 255, 255, 0.87);
+    background-color: rgb(36, 36, 36);
+    border: solid rgb(36, 36, 36);
+    box-shadow: 0px 5px 15px 8px rgba(0, 0, 0, 0.2), 10px 5px 10px 0px rgba(0, 0, 0, 0.19);
+}
+
+.remove-button:hover {
+    color: rgb(0, 0, 0);
+    background-color: rgba(255, 255, 255, 0.87);
+    border: solid rgba(255, 255, 255, 0.87);
+}
+
+.add-trip-input {
+    display: block;
+    margin: 0 auto;
+    width: 20rem;
+}
+
+.add-item-btn {
+    color: rgba(255, 255, 255, 0.87);
+    background-color: rgb(36, 36, 36);
+    border: solid rgb(36, 36, 36);
+    box-shadow: 0px 5px 15px 8px rgba(0, 0, 0, 0.2), 10px 5px 10px 0px rgba(0, 0, 0, 0.19);
+    font-family: 'Pontano Sans', sans-serif;
+    font-size: 17px;
+    width: 12rem;
+    margin: 0 auto;
+    display: block;
+}
+
+.add-item-btn:hover {
+    color: rgb(0, 0, 0);
+    background-color: rgba(255, 255, 255, 0.87);
+    border: solid rgba(255, 255, 255, 0.87);
+    font-family: 'Pontano Sans', sans-serif;
+}
